move : diary 패키지 생성 후 관련 파일 이동
