+++ src/main/java/com/ajou_nice/with_pet/diary/repository/DiaryRepository.java
@@ -1,9 +1,8 @@
-package com.ajou_nice.with_pet.repository;
+package com.ajou_nice.with_pet.diary.repository;
 
-import com.ajou_nice.with_pet.domain.entity.Diary;
+import com.ajou_nice.with_pet.diary.model.entity.Diary;
 import com.ajou_nice.with_pet.dog.model.entity.Dog;
 import com.ajou_nice.with_pet.petsitter.model.entity.PetSitter;
-import com.ajou_nice.with_pet.repository.custom.userdiary.DiaryRepositoryCustom;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
