+++ src/main/java/com/ajou_nice/with_pet/diary/service/UserDiaryService.java
@@ -1,21 +1,23 @@
-package com.ajou_nice.with_pet.service;
+package com.ajou_nice.with_pet.diary.service;
 
-import com.ajou_nice.with_pet.domain.dto.diary.DiaryRequest;
-import com.ajou_nice.with_pet.domain.dto.diary.user.UserDiaryMonthResponse;
-import com.ajou_nice.with_pet.domain.dto.diary.user.UserDiaryResponse;
+import com.ajou_nice.with_pet.diary.model.dto.DiaryRequest;
+import com.ajou_nice.with_pet.diary.model.dto.user.UserDiaryMonthResponse;
+import com.ajou_nice.with_pet.diary.model.dto.user.UserDiaryResponse;
 import com.ajou_nice.with_pet.domain.entity.Category;
 import com.ajou_nice.with_pet.dog.model.entity.Dog;
 import com.ajou_nice.with_pet.domain.entity.Notification;
 import com.ajou_nice.with_pet.domain.entity.User;
-import com.ajou_nice.with_pet.domain.entity.Diary;
+import com.ajou_nice.with_pet.diary.model.entity.Diary;
 import com.ajou_nice.with_pet.domain.entity.UserParty;
 import com.ajou_nice.with_pet.enums.NotificationType;
 import com.ajou_nice.with_pet.exception.AppException;
 import com.ajou_nice.with_pet.exception.ErrorCode;
 import com.ajou_nice.with_pet.dog.repository.DogRepository;
-import com.ajou_nice.with_pet.repository.DiaryRepository;
+import com.ajou_nice.with_pet.diary.repository.DiaryRepository;
 import com.ajou_nice.with_pet.repository.NotificationRepository;
 import com.ajou_nice.with_pet.repository.UserPartyRepository;
+import com.ajou_nice.with_pet.service.NotificationService;
+import com.ajou_nice.with_pet.service.ValidateCollection;
 import java.time.LocalDate;
 import java.time.Period;
 import java.util.List;
