+++ src/main/java/com/ajou_nice/with_pet/dog/model/entity/Dog.java
@@ -3,7 +3,7 @@ package com.ajou_nice.with_pet.dog.model.entity;
 import com.ajou_nice.with_pet.dog.model.dto.update.DogUpdateInfoRequest;
 import com.ajou_nice.with_pet.dog.model.dto.add.DogRegisterRequest;
 import com.ajou_nice.with_pet.domain.entity.BaseEntity;
-import com.ajou_nice.with_pet.domain.entity.Diary;
+import com.ajou_nice.with_pet.diary.model.entity.Diary;
 import com.ajou_nice.with_pet.reservation.model.entity.Reservation;
 import com.ajou_nice.with_pet.enums.DogSize;
 import com.ajou_nice.with_pet.enums.Gender;
