+++ src/test/java/com/ajou_nice/with_pet/diary/DeleteUserDiaryTest.java
@@ -1,7 +1,7 @@
 package com.ajou_nice.with_pet.diary;
 
 import com.ajou_nice.with_pet.domain.entity.Category;
-import com.ajou_nice.with_pet.domain.entity.Diary;
+import com.ajou_nice.with_pet.diary.model.entity.Diary;
 import com.ajou_nice.with_pet.dog.model.entity.Dog;
 import com.ajou_nice.with_pet.group.model.entity.Party;
 import com.ajou_nice.with_pet.domain.entity.User;
@@ -11,12 +11,12 @@ import com.ajou_nice.with_pet.exception.AppException;
 import com.ajou_nice.with_pet.exception.ErrorCode;
 import com.ajou_nice.with_pet.fixture.Fixture;
 import com.ajou_nice.with_pet.repository.CategoryRepository;
-import com.ajou_nice.with_pet.repository.DiaryRepository;
+import com.ajou_nice.with_pet.diary.repository.DiaryRepository;
 import com.ajou_nice.with_pet.dog.repository.DogRepository;
 import com.ajou_nice.with_pet.group.repository.PartyRepository;
 import com.ajou_nice.with_pet.repository.UserPartyRepository;
 import com.ajou_nice.with_pet.repository.UserRepository;
-import com.ajou_nice.with_pet.service.UserDiaryService;
+import com.ajou_nice.with_pet.diary.service.UserDiaryService;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
