+++ src/test/java/com/ajou_nice/with_pet/fixture/Fixture.java
@@ -1,7 +1,7 @@
 package com.ajou_nice.with_pet.fixture;
 
 import com.ajou_nice.with_pet.domain.entity.Category;
-import com.ajou_nice.with_pet.domain.entity.Diary;
+import com.ajou_nice.with_pet.diary.model.entity.Diary;
 import com.ajou_nice.with_pet.dog.model.entity.Dog;
 import com.ajou_nice.with_pet.enums.Gender;
 import com.ajou_nice.with_pet.group.model.entity.Party;
